refactor(tests): Add type annotations.
